Added filesystems and devices managers

+++ kernel/devices/pipedev.c
@@ -9,7 +9,7 @@ pipedev_load(inode_t *inode_unused)
 }
 
 static uint32_t
-pipedev_read(inode_t *inode, uint32_t offset, uint32_t size, void *buf)
+pipedev_read(inode_t *inode, uint64_t offset, uint64_t size, void *buf)
 {
